requirements.txt
python-dotenv
Pillow
openai
streamlit==1.26.0
streamlit
numpy
altair
pandas